+++ src/app/quiz/[quizId]/@recommendation/page.tsx
@@ -1,5 +1,3 @@
-import { Metadata } from 'next';
-
 import { QuizCarousel } from '@/app/quiz/components';
 import { getRecommendationByQuizId } from '@/app/quiz/remotes/recommendation';
 import { Text } from '@/common';
@@ -9,27 +7,6 @@ type Props = {
     quizId: string;
   };
 };
-
-export async function generateMetadata({
-  params: { quizId },
-}: Props): Promise<Metadata> {
-  const quizRecommendModels = await getRecommendationByQuizId(quizId);
-  const isQuizzesExist = quizRecommendModels.length !== 0;
-
-  return {
-    openGraph: {
-      title: isQuizzesExist
-        ? '이런 퀴즈도 있어요!'
-        : '똑스 : 지식을 키우는 첫 시작!',
-      description: isQuizzesExist
-        ? quizRecommendModels[0].quiz.title
-        : '똑스와 함께, 퀴즈로 똑똑해지고 더 나은 습관 만들기',
-      images: '/toks-og.png',
-      type: 'website',
-      siteName: 'Toks',
-    },
-  };
-}
 
 async function RecommendatonPage({ params: { quizId } }: Props) {
   const quizRecommendModels = await getRecommendationByQuizId(quizId);
